refactor(core): split ingest module to satisfy 300-500 line guideline (#15)
The original ingest.ts grew to 923 lines as discovery + slug derivation
+ state/date derivation + glob expansion all landed in one file. Split:

  - ingest.ts (387 lines) — public API, orchestration, idempotency.
  - ingest-derive.ts (382 lines) — slug / state / date / title
    derivation with provenance recording. STATE_ALIASES table lives here.
  - ingest-paths.ts (210 lines) — file collection, recursive walk,
    glob expansion (no runtime dep on a glob library).

Each `derive*` function now takes an explicit input record instead of
threading IngestOptions through, which makes the derivation seam
testable and avoids the option-bag-coupling that grew up while the
file was monolithic.

Also added subpath exports for `@deskwork/core/ingest-derive` and
`@deskwork/core/ingest-paths` so callers that want to reach the
helpers directly (e.g. a future migration tool) can without going
through the orchestrator.

Verified end-to-end via the writingcontrol.org dogfood scenario from
the workplan:
  - 3 essays land in correct lanes (2 Published with their
    datePublished values, 1 Drafting)
  - --apply writes 3 calendar rows + 3 journal records
  - re-running --apply produces only skips (idempotent)

302 tests still green (core 202, cli 61, studio 39).
